Create Interactive_Quiz
This project demonstrates the use of HTML, CSS and JavaScript to perform a task in which a quiz page is created. This interactive Quiz contains general knowledge questions which have multiple-choice answers. The option bar is highlighted in different colour based on if the input is the right answer or not.
Upon completing the quiz, the user is also prompted a choice to restart and re-attempt the quiz